Create abstract class for position classes UAV-286
Everyone should get to work with the position class that they are most
comfortable with. And to enable everyone to get the most out of the
position classes, distance functions should be provided to all classes.
That is what motivates this abstract base class.

The class should have these methods:
laser Distance (that takes another WGS-84 position) -> returns a double
travel distance(that takes another WGS-84 position) -> returns a double
toNvector() -> returns an NVector
toPVector() -> returns a PVector
toLlaCoordinate() -> returns an LlaCoordinate