Add 3 distance tests to abstract position UAV-286
Need to be sure that the abstract position class is working. Add three
tests that make sure the distance from end zone to end zone at nd
stadium is 91.44 meters.

Also change the implementation of distance make it clear to those
reading the source code that distance is returning laser distance.